# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Movieflix API",
        "description": "API responsável por gerenciar filmes.",
        "version": "1.0.0"
    },
    "paths": {
        "/movies": {
            "get": {
                "tags": [
                    "Movie"
                ],
                "summary": "Listar todos os filmes",
                "responses": {
                    "200": {
                        "description": "Lista de filmes retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Movie"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/movies/{id}": {
            "put": {
                "tags": [
                    "Movie"
                ],
                "summary": "Atualizar um filme por ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID do filme a ser atualizado",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MovieInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                    "description": "Filme atualizado com sucesso"
                    },
                    "404": {
                    "description": "Filme não encontrado"
                    },
                    "400": {
                    "description": "Falta dados de entrada obrigatórios"
                    },
                    "500": {
                    "description": "Falha ao atualizar registro do filme"
                    }
                    
                }
            }, "delete":{
                "tags": [
                    "Movie"
                ],
                "summary": "Deletar um filme por ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "ID do filme a ser deletado",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                    "description": "Filme deletado com sucesso"
                    },
                    "404": {
                    "description": "Filme não encontrado"
                    },
                    "500": {
                    "description": "Falha ao deletar registro do filme"
                    }
                    
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Movie": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "ID do filme"
                    },
                    "title": {
                        "type": "string",
                        "description": "Título do filme"
                    },
                    "genre_id": {
                        "type": "integer",
                        "description": "ID do gênero do filme"
                    },
                    "language_id": {
                        "type": "integer",
                        "description": "ID da lingua do filme"
                    },
                    "oscar_count": {
                        "type": "integer",
                        "description": "Quantidade de oscars do filme"
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date",
                        "description": "Data de lançamento do filme"
                    }
                }
            },
            "MovieInput": {
                "type": "object",
                "required": [
                    "title",
                    "genre_id",
                    "language_id",
                    "oscar_count",
                    "release_date"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "Título do filme"
                    },
                    "genre_id": {
                        "type": "integer",
                        "description": "ID do gênero do filme"
                    },
                    "language_id": {
                        "type": "integer",
                        "description": "ID da lingua do filme"
                    },
                    "oscar_count": {
                        "type": "integer",
                        "description": "Quantidade de oscars do filme"
                    },
                    "release_date": {
                        "type": "string",
                        "format": "date",
                        "description": "Data de lançamento do filme"
                    }
                }
            }
        }
    }
}